move jsonschema generation cli test
